Move Greenplum specific headers into greenplum directory.
