+++ src/bin/pg_upgrade/check.c
@@ -13,7 +13,7 @@
 #include "fe_utils/string_utils.h"
 #include "mb/pg_wchar.h"
 #include "pg_upgrade.h"
-
+#include "greenplum/pg_upgrade_greenplum.h"
 
 static void check_new_cluster_is_empty(void);
 static void check_databases_are_compatible(void);
