+++ src/bin/pg_upgrade/greenplum/file_gp.c
@@ -9,10 +9,11 @@
  */
 #include "postgres_fe.h"
 
-#include "pg_upgrade.h"
+#include "pg_upgrade_greenplum.h"
 
 #include "storage/bufpage.h"
 #include "storage/checksum.h"
+
 
 /*
  * rewriteHeapPageWithChecksum
