+++ src/bin/pg_upgrade/greenplum/gpdb4_heap_convert.c
@@ -5,12 +5,13 @@
  */
 #include "postgres_fe.h"
 
-#include "pg_upgrade.h"
+#include "pg_upgrade_greenplum.h"
 
 #include "access/htup_details.h"
 #include "access/xlogdefs.h"
 #include "storage/bufpage.h"
 #include "storage/checksum.h"
+
 
 /*
  * Page format version that we convert to.
