+++ src/bin/pg_upgrade/info.c
@@ -14,6 +14,7 @@
 #include "access/transam.h"
 #include "catalog/pg_class.h"
 
+#include "greenplum/pg_upgrade_greenplum.h"
 
 static void create_rel_filename_map(const char *old_data, const char *new_data,
 						const DbInfo *old_db, const DbInfo *new_db,
