+++ src/bin/pg_upgrade/server.c
@@ -22,6 +22,7 @@
 #include "fe_utils/string_utils.h"
 #include "pg_upgrade.h"
 
+#include "greenplum/pg_upgrade_greenplum.h"
 
 static PGconn *get_db_conn(ClusterInfo *cluster, const char *db_name);
